// IntelliPMS-Frontdesk/src/app/models/revenue-report.model.ts
export class RevenueReportModel {
  date: string;
  roomNumber: number;
  roomTypeId: number;
  roomTypeName: string;
  totalRevenue: number;
  roomRevenue: number;
  roomServiceRevenue: number;
  laundryServiceRevenue: number;
  houseKeepingRevenue: number;
  averageRevenue: number;
  numberCheckedIn: number;
}